relocate functions to navigation context
